requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
click==7.1.2
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
cycler==0.10.0
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
itemadapter==0.1.0
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
matplotlib==3.3.0
mccabe==0.6.1
nltk==3.5
numpy==1.19.1
pandas==1.0.5
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
pyUFbr==0.1.0
queuelib==1.5.0
regex==2020.7.14
scikit-learn==0.23.1
scipy==1.5.2
Scrapy==2.2.1
seaborn==0.10.1
service-identity==18.1.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.48.0
Twisted==20.3.0
w3lib==1.22.0
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.1.0
